refactor: introduce market repo and entity
