+++ koral-earth/app/backend/market/market.repo.ts
@@ -1,0 +1,6 @@
+import { Project } from './market.entity';
+import toucanProjects from '../shared/data/toucanProjects.json';
+
+export const getProjects = async (): Promise<Project[]> => {
+  return toucanProjects.data.projects;
+};
