+++ koral-earth/app/pages/index.tsx
@@ -1,8 +1,7 @@
 import type { GetStaticPropsResult, NextPage, NextPageContext } from 'next';
-import { Project } from '../backend/shared/models/market';
+import { Project } from '../backend/market/market.entity';
 import { Market } from '../frontend/views/market/Market';
-import toucanProjects from '../backend/shared/data/toucanProjects.json';
-import { rowsFromData } from '../frontend/lib/array';
+import { getProjects } from '../backend/market/market.repo';
 
 type Props = {
   projects: Project[];
@@ -13,7 +12,7 @@ const Home: NextPage<Props> = ({ projects }) => <Market projects={projects} />;
 export const getStaticProps = async (
   _context: NextPageContext
 ): Promise<GetStaticPropsResult<Props>> => {
-  const { projects } = toucanProjects.data;
+  const projects = await getProjects();
 
   return {
     props: { projects },
